+++ catlearn/learning_curve/__init__.py
@@ -1,1 +1,1 @@
-from .learning_curve import hierarchy, feature_frequency
+from .learning_curve import hierarchy, feature_frequency, LearningCurve
